fix: queries without args or where
Currently there are cases where it is assumed that the args object is
defined, and also the where object within. This is not always correct,
for example findMany can be used without args or a where.

Update action middleware to ensure operations that can being used
without args or a where object are supported.

Additionally, we are assuming the user is passing valid arguments all
the time, however there are cases where it would be better to allow
Prisma to throw an error rather than trying to modify the params.

Pass root delete operations with no where, findUnique operations with no
where and updateMany operations with no args through to Prisma to throw
a useful error.