fix: returning models for invalid findUnique where
Closes #4

When a findUnique query is passed an invalid where object, for example
one with `id: undefined`, the deleted field is still added to the where
and the action is converted to findFirst. This causes a model to be
incorrectly returned when it should not be, instead the query should
throw due to Prisma validation.

Use the DMMF to validate that the where object has at least one unique
field that is not undefined before changing the action type to
findFirst.