feat: support Prisma v5 find*OrThrow actions (#14)
Prisma v5 moved away from configuring whether to throw or not when a
findOne does not have a result, instead opting to encourage using
findFirstOrThrow and findUniqueOrThrow methods on the client.

The removal of the configurable throwing behaviour meant that the
middleware actions added findFirstOrThrow and findUniqueOrThrow as
options.

Update the list of action middleware to handle orThrow variants and
update the version of Prisma we test against to the latest v5 version.

Also refactor the findUnique middleware to precalculate uniqueFields
and uniqueIndexFields at the top of actionMiddleware, this avoids having
to recalculate when handling each findUnique or findUniqueOrThrow
operation. At the same time it makes sense to refactor the logic for
whether to pass the params through in the findUnique case and when to
throw into separate helpers, this makes it easier to understand when
params are passed through vs when we should throw.