feat: support truthy values for deleted records (#16)
Currently, the createValue function must return a falsy value if the record is not deleted. Change the result filtering helpers so that we can define createValue functions that have a truthy value, such as Date(0), for when the record is not deleted. It should also now be theoretically possible to use a falsy value for when a record is deleted but that does not have test coverage.

Update resultFiltering helpers to use lodash isEqual to compare values
from createValue function. Lodash isEqual does Date comparison in a way
that works for `new Date(0) === new Date(0)`.

Update README with explanation of new behaviour of createValue.

Update nestedReads tests to use Date(0) as the value for non-deleted
records, ensuring we have test coverage for that behaviour.

---------

Co-authored-by: misterapp <misterapptracy@gmail.com>
Co-authored-by: Olivier <olivier.wilkinson@gmail.com>